Use `Long` for anything counting reads (#54)
* Case class hygiene

* Use `Long` for anything counting reads

* Set version to 3.13.0-SNAPSHOT

* Missed a spot

* Stick to `Int` for test data

* Update changelog and manual

+++ src/main/scala/org/broadinstitute/gpp/poolq3/numeric/numeric.scala
@@ -15,10 +15,10 @@ val Log2: Double = math.log(2)
 
 val OneMillion: Double = 1000000.0
 
-def percent(num: Int, denom: Int): Double = if denom == 0 then 0.0 else num * 100.0 / denom
+def percent(num: Long, denom: Long): Double = if denom == 0L then 0.0 else num * 100.0 / denom
 
 def log2(x: Double): Double = math.log(x) / Log2
 
-def logNormalize(num: Int, denom: Int): Double =
-  if denom == 0 then 0
+def logNormalize(num: Long, denom: Long): Double =
+  if denom == 0L then 0L
   else math.log1p(num * OneMillion / denom) / Log2
